+++ apps/hightide-native-storybook/src/User Interaction/MultiSelect.stories.tsx
@@ -28,7 +28,7 @@ export const multiSelect: Story = {
     initialValue: [],
     disabled: false,
     invalid: false,
-    showSearch: true,
+    searchableThreshold: 6,
     readOnly: false,
     placeholder: 'Select…',
     onValueChange: action('onValueChange'),